Move identify to internal
