+++ internal/identify/studio_test.go
